+++ src/layout/dagre.ts
@@ -8,6 +8,7 @@ import dagre from "./dagre/index";
 import { graphlib as IGraphLib } from './dagre/graphlib';
 import { isArray, isNumber, isObject, getEdgeTerminal } from "../util";
 import { Base } from "./base";
+import { PointTuple } from "./types";
 
 type DagreGraph = IGraphLib.Graph;
 
@@ -20,6 +21,9 @@ export class DagreLayout extends Base {
 
   /** 节点对齐方式，可选 UL, UR, DL, DR */
   public align: undefined | "UL" | "UR" | "DL" | "DR";
+
+  /** 布局的起始（左上角）位置 */
+  public begin: PointTuple;
 
   /** 节点大小 */
   public nodeSize: number | number[] | undefined;
@@ -98,7 +102,7 @@ export class DagreLayout extends Base {
    */
   public execute() {
     const self = this;
-    const { nodes, nodeSize, rankdir, combos } = self;
+    const { nodes, nodeSize, rankdir, combos, begin } = self;
     if (!nodes) return;
     const edges = (self.edges as any[]) || [];
     const g = new dagre.graphlib.Graph({
@@ -198,18 +202,38 @@ export class DagreLayout extends Base {
       keepNodeOrder: self.keepNodeOrder,
       nodeOrder: self.nodeOrder,
     });
-    let coord;
+
+    const dBegin = [0, 0];
+    if (begin) {
+      let minX = Infinity;
+      let minY = Infinity;
+      g.nodes().forEach((node: any) => {
+        const coord = g.node(node);
+        if (minX > coord.x) minX = coord.x;
+        if (minY > coord.y) minY = coord.y;
+      });
+      g.edges().forEach((edge: any) => {
+        const coord = g.edge(edge);
+        coord.points.forEach((point: any) => {
+          if (minX > point.x) minX = point.x;
+          if (minY > point.y) minY = point.y;
+        })
+      });
+      dBegin[0] = begin[0] - minX;
+      dBegin[1] = begin[1] - minY;
+    }
+
     g.nodes().forEach((node: any) => {
-      coord = g.node(node);
+      const coord = g.node(node);
       const i = nodes.findIndex((it) => it.id === node);
       if (!nodes[i]) return;
-      nodes[i].x = coord.x;
-      nodes[i].y = coord.y;
+      nodes[i].x = coord.x + dBegin[0];
+      nodes[i].y = coord.y + dBegin[1];
       // @ts-ignore: pass layer order to data for increment layout use
       nodes[i]._order = coord._order;
     });
     g.edges().forEach((edge: any) => {
-      coord = g.edge(edge);
+      const coord = g.edge(edge);
       const i = edges.findIndex((it) => {
         const source = getEdgeTerminal(it, 'source');
         const target = getEdgeTerminal(it, 'target');
@@ -217,6 +241,10 @@ export class DagreLayout extends Base {
       });
       if ((self.edgeLabelSpace) && self.controlPoints && edges[i].type !== "loop") {
         edges[i].controlPoints = coord.points.slice(1, coord.points.length - 1);
+        edges[i].controlPoints.forEach((point: any) => {
+          point.x += dBegin[0];
+          point.y += dBegin[1];
+        })
       }
     });