+++ src/layout/types.ts
@@ -135,6 +135,7 @@ export interface DagreLayoutOptions {
   type: "dagre";
   rankdir?: "TB" | "BT" | "LR" | "RL";
   align?: "UL" | "UR" | "DL" | "DR";
+  begin?: PointTuple;
   nodeSize?: number | number[] | undefined;
   nodesep?: number;
   ranksep?: number;